Changed flatten.go to framework_models.go
