+++ linode/instancenetworking/framework_datasource.go
@@ -2,11 +2,8 @@ package instancenetworking
 
 import (
 	"context"
-	"encoding/json"
 
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
-	"github.com/hashicorp/terraform-plugin-framework/diag"
-	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/linode/linodego"
 	"github.com/linode/terraform-provider-linode/linode/helper"
 )
@@ -17,34 +14,6 @@ func NewDataSource() datasource.DataSource {
 
 type DataSource struct {
 	client *linodego.Client
-}
-
-func (data *DataSourceModel) parseInstanceIPAddressResponse(
-	ctx context.Context, ip *linodego.InstanceIPAddressResponse,
-) diag.Diagnostics {
-	ipv4, diags := flattenIPv4(ctx, ip.IPv4)
-	if diags.HasError() {
-		return diags
-	}
-
-	data.IPV4 = *ipv4
-
-	ipv6, diags := flattenIPv6(ctx, ip.IPv6)
-	if diags.HasError() {
-		return diags
-	}
-
-	data.IPV6 = *ipv6
-
-	id, err := json.Marshal(ip)
-	if err != nil {
-		diags.AddError("Error marshalling json: %s", err.Error())
-		return diags
-	}
-
-	data.ID = types.StringValue(string(id))
-
-	return nil
 }
 
 func (d *DataSource) Configure(
@@ -63,13 +32,6 @@ func (d *DataSource) Configure(
 	}
 
 	d.client = meta.Client
-}
-
-type DataSourceModel struct {
-	LinodeID types.Int64  `tfsdk:"linode_id"`
-	IPV4     types.Object `tfsdk:"ipv4"`
-	IPV6     types.Object `tfsdk:"ipv6"`
-	ID       types.String `tfsdk:"id"`
 }
 
 func (d *DataSource) Metadata(
